Changed contract of the DeserializePacket() method;
DeserializePacket now expects the reader to already be
pointing at the top of the packet object when it is
called. It used to expected the reader to be pointing
immediately before the top packet object. This change
is necessary in order to deserialize a list of packets
(since the caller needs to know whether a token is
the end of an array or the start of another packet)